move some fields to WorkerHeartBeat
